Resolve RemovedInDjango19Warning
RemovedInDjango19Warning: Model class lumberjack.models.LogEntry doesn't declare an explicit app_label and either isn't in an application in INSTALLED_APPS or else was imported before its application was loaded. This will no longer be supported in Django 1.9.

Since the logging is loaded early, it throws this warning although lumberjack is in the `INSTALLED_APPS`.